# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "CarKit API Documentation",
    "version": "0.0.1"
  },
  "host": [
    "carkit.lazyy.fr"
  ],
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/full-tank/{carId}": {
      "post": {
        "tags": [
          "FullTank"
        ],
        "description": "Add a new full tank entry for a car.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Full tank entry data",
            "schema": {
              "type": "object",
              "properties": {
                "quantity": {
                  "example": 50
                },
                "unit": {
                  "example": "liter"
                },
                "cost": {
                  "example": 80.5
                },
                "mileage": {
                  "example": 15000
                },
                "date": {
                  "example": "2023-01-01"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "quantity": 50,
              "unit": "liter",
              "cost": 80.5,
              "mileage": 15000,
              "date": "2023-01-01"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "get": {
        "tags": [
          "FullTank"
        ],
        "description": "Get all full tank entries for a car.",
        "parameters": [
          {
            "in": "path",
            "name": "carId",
            "required": true,
            "description": "Path parameter carId",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "quantity": 50,
                "unit": "liter",
                "cost": 80.5,
                "mileage": 15000,
                "date": "2023-01-01"
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      }
    },
    "/full-tank/{id}": {
      "delete": {
        "tags": [
          "FullTank"
        ],
        "description": "Delete a full tank entry.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Full tank entry deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Full tank entry not found."
            }
          }
        }
      }
    },
    "/car": {
      "post": {
        "tags": [
          "Car"
        ],
        "description": "Create a new car.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Car creation data",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "example": "ice"
                },
                "model": {
                  "example": "Model S"
                },
                "make": {
                  "example": "Tesla"
                },
                "name": {
                  "example": "My Tesla"
                },
                "mileageAtStart": {
                  "example": 10000
                },
                "year": {
                  "example": 2020
                },
                "imageData": {
                  "example": "base64ImageString"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "name": "My Tesla",
              "make": "Tesla",
              "model": "Model S",
              "type": "elec"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "422": {
            "description": "Response",
            "examples": {
              "status": 422,
              "msg": "Required fields missing."
            }
          }
        }
      },
      "get": {
        "tags": [
          "Car"
        ],
        "description": "Get all cars for the connected user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response",
            "examples": [
              {
                "id": "uuid",
                "name": "My Tesla",
                "make": "Tesla",
                "model": "Model S",
                "type": "elec"
              }
            ]
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          }
        }
      }
    },
    "/car/{id}": {
      "get": {
        "tags": [
          "Car"
        ],
        "description": "Get car details by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "name": "My Tesla",
              "make": "Tesla",
              "model": "Model S",
              "type": "elec"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "put": {
        "tags": [
          "Car"
        ],
        "description": "Update car details.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Car update data",
            "schema": {
              "type": "object",
              "properties": {
                "make": {
                  "example": "Tesla"
                },
                "model": {
                  "example": "Model S"
                },
                "name": {
                  "example": "My Tesla"
                },
                "mileageAtStart": {
                  "example": 10000
                },
                "year": {
                  "example": 2020
                },
                "type": {
                  "example": "elec"
                },
                "imageData": {
                  "example": "base64ImageString"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "id": "uuid",
              "name": "Updated Tesla",
              "make": "Tesla",
              "model": "Model S",
              "type": "elec"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Car"
        ],
        "description": "Delete a car.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Path parameter id",
            "schema": {
              "type": "string",
              "example": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "Car deleted successfully."
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "Car not found."
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
}